1.5.2

+++ src/net/tnemc/core/economy/currency/Tier.java
@@ -1,5 +1,7 @@
 package net.tnemc.core.economy.currency;
 
+import java.math.BigDecimal;
+
 /**
  * Created by creatorfromhell on 8/9/2017.
  * All rights reserved.
@@ -24,5 +26,5 @@ public interface Tier {
   /**
    * @return Gets the weight of this tier. A weight of one is equivalent to 1.0 of currency.
    */
-  double weight();
+  BigDecimal weight();
 }
